Renamed SmallSubjectImage to SubjectImage
